Fix screenstack not popping properly.
When `Stack::pop` was called, a copy of the newly-revealed screen would
be added to the stack. When it was called again, the active screen
would be destroyed, but then replaced with itself - resulting in just
showing a blank screen.

Fix the behaviour by not pushing the screen on to the stack a second time.